a couple of fixes

+++ controller/controller_test.go
@@ -77,7 +77,7 @@ func TestController_GameCRUD(t *testing.T) {
 		game, err := client.Status(ctx, &pb.StatusRequest{ID: gameID})
 		require.Nil(t, err)
 		require.Equal(t, "running", game.Game.Status)
-		require.Equal(t, int64(1000), game.Game.SnakeTimeout)
+		require.Equal(t, int32(1000), game.Game.SnakeTimeout)
 		require.Equal(t, "multi-player", game.Game.Mode)
 		require.NotNil(t, game.LastFrame)
 		require.Equal(t, int32(0), game.LastFrame.Turn)
